requirements.txt
pandas
sentence-transformers
langdetect
torch
scikit-learn
tqdm
openai
google-generativeai
python-dotenv
Flask
waitress